feat(interactive): adds configuration.xml file override
Adds two new configuration values (`configurationXml` and `configurationXmlConfigMap`) that can be used to replace the configuration.xml file provided by the image.

XML content can be specified inline via the `configurationXml` value and a ConfigMap will be created (if it does not already exist). If a name (ConfigMap) is provided via the `configurationXmlConfigMap` value, the existing ConfigMap will be used instead.

Ref: #6